Split Cause and Context
Any event now has two objects attached to it, a cause and a context. The cause contains
an ordered stack of objects which are the direct cause of the event. The context contains
objects that are tangentially related to the event being thrown that the thrower of the
event thinks would be useful for proper handling of the event.

In the cause only the order matters, in the context only the names matter. A cause has
no names and a context has no ordering.

Previous the values of these two objects were both present in the cause which can cause
some confusion and misaction. Forexample you could attach a player to an event to which
there are only slightly related to but if no other player is in the cause before them then
cause.first(Player.class) will still return that player even though there are by no means
the cause of the event.